fix: Resolve duplicate clipPath URL on multiple icon instances
The origin of this issue is using a clipPath URL inside an instance
of an icon. Those URLs should be unique over the entire webpage
as they reference the first element found with the ID.

This results in datatable filters as a reference to the first icon. So the clear-icons are only visible if the first one is as well.