Explain the default behavior of primevue.components.exclude in the Nuxt module configuration and how to mitigate.
In the documentation for configuring the Nuxt module components, there is no indication that Chart and Editor are excluded by default. The only way to find out is if you happen upon a single issue in the github repo of the nuxt module npm package or read the full source code of the module. Although the intention of this default behavior is to avoid compatibility issues because of the dependencies of those 2 components, it leads to even more confusion when you know what they need but they aren't being detected as components. They are excluded even if you use the * option in include. This change introduces a helpful tip into the docs page so users know what to do when they want to use those components.