Fix ConfirmPopup placement on stopped events
The previous implementation only sets the ConfirmPopup position in
bindOutsideClickListener, when a click is registered on the root
window that isn't on the popup itself.

This is an issue, as it does not work if the click event that caused
the popup to appear does not propagate up to the root window.

Instead, this change always positions the window when it appears.
This is already what the Popover component does, which does not exhibits
the same bug.

Fix #5929