Fix #6371: Add node and expanded data on nodetoggleicon.
Nodetoggleicon slot is used to customize expand / collapse and loading
icon, but when loading mode is set to icon mode and a node is loading,
TreeNode use nodetoggleicon without any props instead of binding node and
expanded.
This commit bind node and expanded props to nodetoggleicon slot when
loading mode is set to icon and node.loading is true.