fix(date-picker): inline used together with range and time options
When using selectionMode = range, showTime and inline options
together the component would not allow the time of the second
date to be edited.

This seems like a bug to me rather than a intended
behavior of the component.

If it is for some reason intended, it should be documented
that this won't indeed work.