Ensure canvas exists when initializing Chart.js
When initializing the chart, the import for chart.js is dynamic which
means its possible for the canvas to be removed before we call
chart.js's initialization function.  This can cause sporadic failures
during tests if components are being loaded for short periods of times.
During general use it's unlikely to happen as the import is likely to
be resolved quickly.